part2 - phonebook - step9 completed

+++ part-2/phonebook/src/App.js
@@ -57,13 +57,15 @@ const App = () => {
   }
 
   const handleDelete = (id) => {
-    console.log(id)
-    servicePerson
-      .deletePerson(id)
-      .then(response => {
-        setPersons(persons.filter(person => person.id !== id
-        ))
-      })
+    const toDelete = persons.find(person => person.id === id)
+    if (window.confirm(`Delete ${toDelete.name}?`)) {
+      servicePerson
+        .deletePerson(id)
+        .then(response => {
+          setPersons(persons.filter(person => person.id !== id
+          ))
+        })
+    }
   }
 
   return (
